refactor: add Repository class
