+++ Program.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
+using LabManager.Repositories;
 
 var databaseSetup = new DatabaseSetup();
+var computerRepository = new ComputerRepository();
 
 var modelName = args[0];
 var modelAction = args[1];
@@ -10,21 +12,10 @@ if(modelName == "Computer")
 {
     if(modelAction == "List")
     {
-        Console.WriteLine("List Computer");
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers";
-
-        var reader = command.ExecuteReader();
-
-        while(reader.Read())
+        foreach (var computer in computerRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2));
+            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
         }
-
-        connection.Close();
     }
 
     if(modelAction == "New")
